Start working on dungeon generation
!!!Everything here is untested!!!

MoveRooms(int direction) functionality added
RelativeOffset(int dir) returns the Vector2 offset of an int direction.
InBounds(Vector2 coord) returns whether a coordinate is in bounds of the map.
IsContinuous(Vector2 a, Vector2 b) uses flood filling to try and find a walk-able path between a and b

OnMap() returns "?" and " " correctly
Add a new constructor to Vector2() so it defaults to 0,0
equals(Object o) overridden for vector2 stuff
IsEqual(Vector2 a, Vector2 b) for above
Sum(Vector2 a, Vector2 b) for adding two vectors together
Product(Vector2 a, int b) for multiplying a vector by an integer.

Add a Vector2 position for the player (and respective getter/setter)