Basic Dungeon Navigation
Add Status to determine "where" the player is in the game
Better initial dungeon creation, will regenerate when completing the dungeon floor.
Future setup for the rest of the game is present.
Add a shorthand for GetRoom(Player.GetPos)
Add PlayerMovePrompt() to ask for user input (either numerical or cardinal) and try to move. It will keep asking until the input and move are legal, then calls MoveTo()
Tidy up MoveTo() handling the door and key so it actually works as intended.
Fix RelativeOffset(), the vertical axis was flipped.
Add a couple more IntInput() overrides for asking for input after displaying a message.
Removes the key from a room when visiting a key with a room.
Add ClearEnemy() for later.
Add a Vector2.toString() override