Add enemy population and basic combat
Still a lot to flush out, especially user-friendliness related, but the foundation is there.

Added a random parameter to any function that uses it to normalize seeds so a consistent dungeon/enemy layout can be generated.
Dungeons now get progressively more difficult and larger the deeper the player goes.
Combat works by forcing the player to either kill or be killed by the enemy in the room.
Enemies will populate the dungeon floor, increasing in amount.
The code will exit when the player dies
Fix the math challenge never succeeding
Enemies will generate with a random name from a list of names. This has no effect on gameplay.